Add validating whether new taxonomy is in NCBI in search_taxonomies function. If so, add this new one in local database. Then we can move acceptNewTaxonomy from mavedb-ui and solve autocomplete problem.
