+++ src/mavedb/lib/identifiers.py
@@ -1,7 +1,5 @@
-import httpx
 import os
 from datetime import date
-from fastapi.exceptions import HTTPException
 from typing import Optional, Union
 
 import eutils
@@ -21,7 +19,6 @@ from mavedb.models.raw_read_identifier import RawReadIdentifier
 from mavedb.models.refseq_identifier import RefseqIdentifier
 from mavedb.models.refseq_offset import RefseqOffset
 from mavedb.models.target_gene import TargetGene
-from mavedb.models.taxonomy import Taxonomy
 from mavedb.models.uniprot_identifier import UniprotIdentifier
 from mavedb.models.uniprot_offset import UniprotOffset
 
@@ -415,38 +412,3 @@ async def create_external_gene_identifier_offset(
 
     identifier_offset_attribute = EXTERNAL_GENE_IDENTIFIER_OFFSET_ATTRIBUTES[external_gene_identifier.db_name]
     setattr(target_gene, identifier_offset_attribute, external_gene_identifier_offset)
-
-async def find_or_create_taxonomy(db: Session, taxonomy: Taxonomy):
-    """
-    Find an existing taxonomy ID record with the specified tax_id int, or create a new one.
-
-    :param db: An active database session
-    :param tax_id: A valid taxonomy ID
-    :return: An existing Taxonomy containing the specified taxonomy ID, or a new, unsaved Taxonomy
-    """
-    taxonomy_record = db.query(Taxonomy).filter(Taxonomy.tax_id == taxonomy.tax_id).one_or_none()
-    if not taxonomy_record:
-        url = f"https://api.ncbi.nlm.nih.gov/datasets/v2alpha/taxonomy/taxon/{taxonomy.tax_id}"
-        async with httpx.AsyncClient() as client:
-            response = await client.get(url)
-            if response.status_code == 200:
-                data = response.json()
-                # Process the retrieved data as needed
-                ncbi_taxonomy = data['taxonomy_nodes'][0]['taxonomy']
-                ncbi_taxonomy.setdefault('organism_name', 'NULL')
-                ncbi_taxonomy.setdefault('common_name', 'NULL')
-                ncbi_taxonomy.setdefault('rank', 'NULL')
-                ncbi_taxonomy.setdefault('has_described_species_name', False)
-                taxonomy_record = Taxonomy(tax_id=ncbi_taxonomy['tax_id'],
-                                           organism_name=ncbi_taxonomy['organism_name'],
-                                           common_name=ncbi_taxonomy['common_name'],
-                                           rank=ncbi_taxonomy['rank'],
-                                           has_described_species_name=ncbi_taxonomy['has_described_species_name'],
-                                           url="https://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?mode=info&id="+str(ncbi_taxonomy['tax_id']),
-                                           article_reference="NCBI:txid"+str(ncbi_taxonomy['tax_id']))
-                db.add(taxonomy_record)
-                db.commit()
-                db.refresh(taxonomy_record)
-            else:
-                raise HTTPException(status_code=404, detail=f"Taxonomy with taxID {taxonomy.tax_id} not found")
-    return taxonomy_record
